♻️ Extract layer, capture, and runs into bin/cli modules
Continue the CLI refactor by moving layer, capture, and runs commands
into dedicated bin/cli modules and routing through withDaemon. Shrinks
lattices.ts by ~480 lines while keeping the thin router pattern.
